% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_classif.R
\name{predict_classif}
\alias{predict_classif}
\title{Predcit function for classif tasks made in Python}
\usage{
predict_classif(model, newdata)
}
\arguments{
\item{model}{- trained model that predcitions we are look for}

\item{newdata}{- test dataset}
}
\description{
Predcit function for classif tasks made in Python
}
