% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_regr.R
\name{predict_regr}
\alias{predict_regr}
\title{Predcit function for regr tasks made in Python}
\usage{
predict_regr(model, newdata)
}
\arguments{
\item{model}{- trained model that predcitions we are look for}

\item{newdata}{- test dataset}
}
\description{
Predcit function for regr tasks made in Python
}
