% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_explain.R
\name{py_explain}
\alias{py_explain}
\title{DALEX is powerfull model to explain black-box models. Unfortuantely this tool is dedicated for R. This function creates
explainer object from DALEX package using pickle python model}
\usage{
py_explain(path, task, data = NULL, y = NULL,
  label = tail(class(model), 1))
}
\arguments{
\item{path}{- a path to pickle file}

\item{task}{- specify task, "regr" or "classif"}

\item{data}{- newdata}

\item{y}{- responses for newdata}

\item{label}{- label}
}
\description{
DALEX is powerfull model to explain black-box models. Unfortuantely this tool is dedicated for R. This function creates
explainer object from DALEX package using pickle python model
}
\examples{
library(PyWrapper)
explainer <- py_explain("pickle.pkl", "classif", titanic_test, y, "Titanic explainer")
DALEX::model_performance(explainer)

}
